refactor(settings): decompose settings modal into composable modules
