+++ apps/desktop/src/components/features/settings/index.ts
@@ -10,7 +10,9 @@ export {
   removePromptOverride,
   resetPromptOverrideToBuiltin,
   selectedModelKeyForRole,
+  togglePromptOverrideEnabled,
   toRoleVariantOptions,
+  updatePromptOverrideTemplate,
   updateRoleDefault,
 } from "./settings-modal-model";
 export {
