+++ apps/desktop/src/components/features/settings/settings-modal-content.tsx
@@ -1,0 +1,197 @@
+import type { ReactElement } from "react";
+import { GeneralSettingsSection } from "./settings-general-section";
+import type {
+  PromptRoleTabId,
+  RepositorySectionId,
+  SettingsSectionId,
+} from "./settings-modal-constants";
+import { resolveInheritedPromptPreview } from "./settings-modal-normalization";
+import { RepositorySidebar } from "./settings-modal-sidebars";
+import { PromptOverridesSection } from "./settings-prompt-overrides-section";
+import { RepositoryAgentsSection } from "./settings-repository-agents-section";
+import { RepositoryConfigurationSection } from "./settings-repository-configuration-section";
+import type { SettingsModalController } from "./use-settings-modal-controller";
+
+type SettingsModalContentProps = {
+  section: SettingsSectionId;
+  repositorySection: RepositorySectionId;
+  globalPromptRoleTab: PromptRoleTabId;
+  repoPromptRoleTab: PromptRoleTabId;
+  isInteractionDisabled: boolean;
+  controller: SettingsModalController;
+  onRepositorySectionChange: (next: RepositorySectionId) => void;
+  onGlobalPromptRoleTabChange: (next: PromptRoleTabId) => void;
+  onRepoPromptRoleTabChange: (next: PromptRoleTabId) => void;
+};
+
+export function SettingsModalContent({
+  section,
+  repositorySection,
+  globalPromptRoleTab,
+  repoPromptRoleTab,
+  isInteractionDisabled,
+  controller,
+  onRepositorySectionChange,
+  onGlobalPromptRoleTabChange,
+  onRepoPromptRoleTabChange,
+}: SettingsModalContentProps): ReactElement {
+  const {
+    isLoadingSettings,
+    isLoadingCatalog,
+    isSaving,
+    isPickingWorktreeBasePath,
+    settingsError,
+    catalogError,
+    snapshotDraft,
+    catalog,
+    repoPaths,
+    selectedRepoPath,
+    selectedRepoConfig,
+    selectedRepoBranches,
+    isLoadingSelectedRepoBranches,
+    selectedRepoBranchesError,
+    promptValidationState,
+    selectedRepoPromptValidationErrors,
+    selectedRepoPromptValidationErrorCount,
+    globalPromptRoleTabErrorCounts,
+    selectedRepoPromptRoleTabErrorCounts,
+    modelOptions,
+    agentOptions,
+    modelGroups,
+    setSelectedRepoPath,
+    retrySelectedRepoBranchesLoad,
+    updateSelectedRepoConfig,
+    updateGlobalPromptOverrides,
+    updateRepoPromptOverrides,
+    updateSelectedRepoAgentDefault,
+    clearSelectedRepoAgentDefault,
+    pickWorktreeBasePath,
+  } = controller;
+
+  if (settingsError) {
+    return (
+      <div className="rounded-md border border-destructive/40 bg-destructive/10 p-3 text-sm text-destructive">
+        Failed to load settings: {settingsError}
+      </div>
+    );
+  }
+
+  if (isLoadingSettings || !snapshotDraft) {
+    return (
+      <div className="rounded-md border border-border bg-muted/50 p-4 text-sm text-muted-foreground">
+        Loading settings...
+      </div>
+    );
+  }
+
+  if (section === "general") {
+    return <GeneralSettingsSection />;
+  }
+
+  if (section === "prompts") {
+    return (
+      <PromptOverridesSection
+        title="Global Prompt Overrides"
+        description="Global overrides apply to every repository unless a repository-specific enabled override exists for the same prompt."
+        tab={globalPromptRoleTab}
+        errorCountsByTab={globalPromptRoleTabErrorCounts}
+        overrides={snapshotDraft.globalPromptOverrides}
+        validationErrors={promptValidationState.globalErrors}
+        disabled={isInteractionDisabled}
+        onTabChange={onGlobalPromptRoleTabChange}
+        onUpdateOverrides={updateGlobalPromptOverrides}
+        resolveInheritedPreview={(_templateId, builtinTemplate, override) =>
+          override && override.enabled !== false
+            ? undefined
+            : {
+                sourceLabel: "Builtin prompt",
+                template: builtinTemplate,
+              }
+        }
+      />
+    );
+  }
+
+  return (
+    <div className="grid h-full lg:grid-cols-[240px_minmax(0,1fr)]">
+      <RepositorySidebar
+        repoPaths={repoPaths}
+        selectedRepoPath={selectedRepoPath}
+        selectedRepositorySection={repositorySection}
+        disabled={isInteractionDisabled}
+        selectedRepoPromptValidationErrorCount={selectedRepoPromptValidationErrorCount}
+        repoPromptErrorCountByPath={promptValidationState.repoErrorCountByPath}
+        onSelectRepoPath={setSelectedRepoPath}
+        onSelectSection={onRepositorySectionChange}
+      />
+
+      <div className="min-w-0 space-y-4">
+        {repoPaths.length === 0 ? (
+          <div className="rounded-md border border-warning-border bg-warning-surface p-3 text-sm text-warning-surface-foreground">
+            Add a repository first, then configure repository settings.
+          </div>
+        ) : null}
+
+        {repositorySection === "configuration" ? (
+          <RepositoryConfigurationSection
+            selectedRepoConfig={selectedRepoConfig}
+            selectedRepoBranches={selectedRepoBranches}
+            selectedRepoBranchesError={selectedRepoBranchesError}
+            isLoadingSettings={isLoadingSettings}
+            isSaving={isSaving}
+            isPickingWorktreeBasePath={isPickingWorktreeBasePath}
+            isLoadingSelectedRepoBranches={isLoadingSelectedRepoBranches}
+            onRetrySelectedRepoBranchesLoad={retrySelectedRepoBranchesLoad}
+            onPickWorktreeBasePath={pickWorktreeBasePath}
+            onUpdateSelectedRepoConfig={updateSelectedRepoConfig}
+          />
+        ) : null}
+
+        {repositorySection === "agents" ? (
+          <RepositoryAgentsSection
+            selectedRepoConfig={selectedRepoConfig}
+            isLoadingCatalog={isLoadingCatalog}
+            isLoadingSettings={isLoadingSettings}
+            isSaving={isSaving}
+            catalogError={catalogError}
+            catalog={catalog}
+            modelOptions={modelOptions}
+            agentOptions={agentOptions}
+            modelGroups={modelGroups}
+            onUpdateSelectedRepoConfig={updateSelectedRepoConfig}
+            onUpdateSelectedRepoAgentDefault={updateSelectedRepoAgentDefault}
+            onClearSelectedRepoAgentDefault={clearSelectedRepoAgentDefault}
+          />
+        ) : null}
+
+        {repositorySection === "prompts" ? (
+          selectedRepoConfig ? (
+            <PromptOverridesSection
+              title="Repository Prompt Overrides"
+              description="Repository overrides take precedence over global overrides when enabled."
+              tab={repoPromptRoleTab}
+              errorCountsByTab={selectedRepoPromptRoleTabErrorCounts}
+              overrides={selectedRepoConfig.promptOverrides}
+              validationErrors={selectedRepoPromptValidationErrors}
+              disabled={isInteractionDisabled}
+              onTabChange={onRepoPromptRoleTabChange}
+              onUpdateOverrides={updateRepoPromptOverrides}
+              resolveInheritedPreview={(templateId, builtinTemplate, repoOverride) =>
+                resolveInheritedPromptPreview(
+                  templateId,
+                  repoOverride,
+                  snapshotDraft.globalPromptOverrides,
+                  builtinTemplate,
+                )
+              }
+            />
+          ) : (
+            <div className="rounded-md border border-warning-border bg-warning-surface p-3 text-sm text-warning-surface-foreground">
+              Select a repository to configure repository-level prompts.
+            </div>
+          )
+        ) : null}
+      </div>
+    </div>
+  );
+}
